Third commit:Modified:main.js, New file: utils/functions.js(development-branch)
